+++ src/middlewares/eth2Cfx.js
@@ -2,11 +2,11 @@ const { createScaffoldMiddleware } = require('json-rpc-engine');
 const { Conflux, Transaction, sign } = require('js-conflux-sdk');
 const _ = require('lodash');
 const { ethers } = require('ethers');
+const { RLP } = require("ethers/lib/utils");
 const format = require('../utils/format');
 const { numToHex, delKeys, createAsyncMiddleware } = require('../utils');
 const adaptErrorMsg = require('../utils/adaptErrorMsg')
 const ethRawTxConverter = require('../utils/ethRawTxConverter');
-const { RLP } = require("ethers/lib/utils");
 
 const defaultOptions = {
   respAddressBeHex: false,
@@ -59,7 +59,7 @@ function cfx2Eth(options = defaultOptions) {
   // }
 
   async function getNetworkId() {
-    networkId = networkId || (await cfx.getStatus()).networkId
+    networkId = (networkId || (await cfx.getStatus()).networkId)
     return networkId
   }
 
@@ -108,13 +108,14 @@ function cfx2Eth(options = defaultOptions) {
   }
 
   async function call(req, res, next) {
+    await formatEpochOfParams(req.params, 1);
     format.formatCommonInput(req.params, await getNetworkId());
     await next();
   }
 
   async function estimateGas(req, res, next) {
-
-    req.params = format.formatCommonInput(req.params, await getNetworkId());
+    await formatEpochOfParams(req.params, 1);
+    format.formatCommonInput(req.params, await getNetworkId());
     await next();
     if (!res.result) return;
     res.result = res.result.gasUsed;
@@ -129,31 +130,43 @@ function cfx2Eth(options = defaultOptions) {
   async function getBlockByHash(req, res, next) {
     await next();
     if (!res.result) return;
-    format.formatBlock(res.result);
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber, req.params[1]);
+    res.result = pivotBlock;
+    format.formatBlock(res.result, await getNetworkId(), respAddressBeHex);
   }
 
   async function getBlockByNumber(req, res, next) {
     format.formatEpochOfParams(req.params, 0);
     await next();
     if (!res.result) return;
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber, req.params[1]);
+    res.result = pivotBlock;
     format.formatBlock(res.result, await getNetworkId(), respAddressBeHex);
   }
 
   async function getBlockTransactionCountByHash(req, res, next) {
     await next();
     if (!res.result) return;
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber);
+    res.result = pivotBlock;
     res.result = numToHex(res.result.transactions.length);
   }
 
   async function getBlockTransactionCountByNumber(req, res, next) {
     await next();
     if (!res.result) return;
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber);
+    res.result = pivotBlock;
     res.result = util.numToHex(res.result.transactions.length);
   }
 
   async function getCode(req, res, next) {
     req.params[0] = await formatAddress(req.params[0]);
-    format.formatEpochOfParams(req.params, 1);
+    await formatEpochOfParams(req.params, 1);
     // console.log("getcode formated params:", req.params)
     await next();
     // console.log("getcode next response:", res)
@@ -179,7 +192,7 @@ function cfx2Eth(options = defaultOptions) {
   async function getStorageAt(req, res, next) {
     req.params[0] = await formatAddress(req.params[0]);
     req.params[1] = ethers.utils.hexZeroPad(req.params[1], 32);
-    format.formatEpochOfParams(req.params, 2);
+    await formatEpochOfParams(req.params, 2);
     await next();
     res.result = res.result || "0x"
   }
@@ -189,6 +202,9 @@ function cfx2Eth(options = defaultOptions) {
     req.params[1] = true;
     await next();
     if (!res.result) return;
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber);
+    res.result = pivotBlock;
     res.result = res.result.transactions[index];
     await _formatTxAndFeedBlockNumber(res.result);
   }
@@ -198,6 +214,9 @@ function cfx2Eth(options = defaultOptions) {
     req.params[1] = true;
     await next();
     if (!res.result) return;
+    let epochNumber = res.result.epochNumber;
+    let pivotBlock = await _getEpochAsBlock(epochNumber);
+    res.result = pivotBlock;
     res.result = res.result.transactions[index];
     await _formatTxAndFeedBlockNumber(res.result);
   }
@@ -212,15 +231,12 @@ function cfx2Eth(options = defaultOptions) {
     req.params[0] = await formatAddress(req.params[0], respAddressBeHex);
     const isPending = req.params[1] === 'pending';
     if (isPending) {
-      req.method = 'cfx_getAccountPendingInfo';
+      req.method = 'txpool_nextNonce';
       req.params = req.params.slice(0, 1);
     } else {
-      format.formatEpochOfParams(req.params, 1);
+      await formatEpochOfParams(req.params, 1);
     }
     await next();
-    if (isPending) {
-      res.result = res.result.localNonce;
-    }
   }
 
   async function getTransactionReceipt(req, res, next) {
@@ -301,7 +317,7 @@ function cfx2Eth(options = defaultOptions) {
 
   async function getBalance(req, res, next) {
     req.params[0] = await formatAddress(req.params[0]);
-    format.formatEpochOfParams(req.params, 1);
+    await formatEpochOfParams(req.params, 1);
     await next();
   }
 
@@ -343,8 +359,8 @@ function cfx2Eth(options = defaultOptions) {
   async function _formatFilter(filter) {
     let { fromBlock, toBlock, blockHash, blockHashes, address } = filter
 
-    filter.fromEpoch = format.formatEpoch(fromBlock);
-    filter.toEpoch = format.formatEpoch(toBlock);
+    filter.fromEpoch = format.formatEpoch(fromBlock); // TODO support eip1898
+    filter.toEpoch = format.formatEpoch(toBlock); // TODO support eip1898
 
     if (blockHash) {
       if (_.isArray(blockHashes)) {
@@ -367,6 +383,49 @@ function cfx2Eth(options = defaultOptions) {
 
   async function formatAddress(address, toHex) {
     return format.formatAddress(address, await getNetworkId(), toHex)
+  }
+
+  async function formatEpochOfParams(params, index) {
+    /*
+      process EIP-1898 https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md
+      { "blockNumber": "0x0" }
+      { "blockHash": "0xd4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3" }
+    */
+    if (params[index] && typeof params[index] === 'object') {
+      if (params[index].blockHash) {
+        const block = await cfx.getBlockByHash(params[index].blockHash);
+        if (!block) throw new Error('block not found');
+        params[index] = block.epochNumber;
+      }
+      if (params[index].blockNumber) {
+        params[index] = params[index].blockNumber;
+      }
+    }
+    
+    format.formatEpochOfParams(req.params, 1);
+  }
+
+  // NOTE: gasLimit, gasUsed, size, transactionsRoot and other fields is not adapted, direct use the pivot block
+  async function _getEpochAsBlock(epochNumber, includeTx = false) {
+    let blockHashes = await cfx.getBlocksByEpoch(epochNumber);
+    if (blockHashes.length === 0) return null;
+    let txes = [];
+    let blocks = [];
+    // TODO: use batch rpc call
+    for(let hash of blockHashes) {
+      let block = await cfx.getBlockByHash(hash, true);
+      if (block) {
+        blocks.push(block);
+        txes = txes.concat(block.transactions);
+      }
+    }
+    if (blocks.length === 0) return null;
+    let pivotBlock = blocks[blocks.length - 1];
+    pivotBlock.transactions = txes.filter(tx => tx.status === 0);
+    if (!includeTx) {
+      pivotBlock.transactions = pivotBlock.transactions.map(tx => tx.hash);
+    }
+    return pivotBlock;
   }
 
   function isCfxTransaction(rawTransaction) {