+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v-next
 
-1. Adapt `getBlockByNumber` to include all tx from one Epoch
+1. Adapt block related methods to include all tx from one Epoch
+2. Support EIP-1898
 
 ## v0.5.0